Continue modernization.
Much of what this package does can now be done by functions in the Python standard library, which did not exist in 2011. In 2024, for example, there is no need whatsoever to use a regular expression to parse a timestamp.

For the time being, this package will continue to exist as a thin wrapper around built-in functionality. In the future it may become entirely redundant.